+++ packages/frontend/src/api.test.ts
@@ -1,44 +1,42 @@
-import { beforeEach, describe, expect, it, vi } from 'vitest';
-import createFetchMock from 'vitest-fetch-mock';
-import { lightdashApi } from './api';
-
-const fetchMocker = createFetchMock(vi);
-
-fetchMocker.enableMocks();
+import nock from 'nock';
+import { describe, expect, it } from 'vitest';
+import { BASE_API_URL, lightdashApi } from './api';
 
 describe('api', () => {
-    beforeEach(() => {
-        vi.resetAllMocks();
-        fetchMocker.doMock();
-        fetchMocker.mockResponse(async () => ({
-            body: JSON.stringify({
+    it('should handle success response', async () => {
+        const scope = nock(BASE_API_URL)
+            .matchHeader('Content-Type', 'application/json')
+            .matchHeader('Lightdash-Request-Method', 'WEB_APP')
+            .get('/api/v1/test')
+            .reply(200, {
                 status: 'ok',
                 results: 'test',
-            }),
-        }));
-    });
+            });
 
-    it('should handle success response', async () => {
         const result = await lightdashApi({
             method: 'GET',
             url: '/test',
             body: null,
             headers: undefined,
         });
+
+        scope.done();
+
         expect(result).toEqual('test');
-        expect(fetchMocker).toHaveBeenCalledTimes(1);
-        expect(fetchMocker).toHaveBeenCalledWith('/api/v1/test', {
-            method: 'GET',
-            headers: {
-                'Content-Type': 'application/json',
-                'Lightdash-Request-Method': 'WEB_APP',
-            },
-            body: null,
-        });
+        expect(scope.isDone()).toBe(true);
     });
 
     it('should allow custom headers', async () => {
-        await lightdashApi({
+        const scope = nock(BASE_API_URL)
+            .matchHeader('Content-Type', 'application/json')
+            .matchHeader('Lightdash-Request-Method', 'TEST')
+            .get('/api/v1/test')
+            .reply(200, {
+                status: 'ok',
+                results: 'another test',
+            });
+
+        const result = await lightdashApi({
             method: 'GET',
             url: '/test',
             body: null,
@@ -46,14 +44,10 @@ describe('api', () => {
                 'Lightdash-Request-Method': 'TEST',
             },
         });
-        expect(fetchMocker).toHaveBeenCalledTimes(1);
-        expect(fetchMocker).toHaveBeenCalledWith('/api/v1/test', {
-            method: 'GET',
-            headers: {
-                'Content-Type': 'application/json',
-                'Lightdash-Request-Method': 'TEST',
-            },
-            body: null,
-        });
+
+        scope.done();
+
+        expect(scope.isDone()).toBe(true);
+        expect(result).toEqual('another test');
     });
 });