+++ packages/frontend/src/components/Explorer/CustomMetricModal/index.tsx
@@ -1,4 +1,5 @@
 import {
+    AdditionalMetric,
     canApplyFormattingToCustomMetric,
     CustomFormat,
     CustomFormatType,
@@ -10,7 +11,6 @@ import {
     MetricType,
     NumberSeparator,
 } from '@lightdash/common';
-
 import {
     Accordion,
     Button,
@@ -22,8 +22,7 @@ import {
     Title,
 } from '@mantine/core';
 import { useForm } from '@mantine/form';
-import { useFeatureFlagEnabled } from 'posthog-js/react';
-import { useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import { ValueOf } from 'type-fest';
 import { v4 as uuidv4 } from 'uuid';
 import { FormatForm } from '../../../features/tableCalculation/components/FormatForm';
@@ -46,10 +45,6 @@ export const CustomMetricModal = () => {
         item,
         type: customMetricType,
     } = useExplorerContext((context) => context.state.modals.additionalMetric);
-
-    const isCustomMetricFormattingEnabled = useFeatureFlagEnabled(
-        'custom-metrics-formatting',
-    );
 
     const toggleModal = useExplorerContext(
         (context) => context.actions.toggleAdditionalMetricModal,
@@ -79,7 +74,6 @@ export const CustomMetricModal = () => {
     }
 
     const canApplyFormatting =
-        isCustomMetricFormattingEnabled &&
         dimensionToCheck &&
         customMetricType &&
         canApplyFormattingToCustomMetric(dimensionToCheck, customMetricType);
@@ -91,7 +85,12 @@ export const CustomMetricModal = () => {
 
     const { projectUuid, fieldsMap, startOfWeek } = useDataForFiltersProvider();
 
-    const form = useForm({
+    const form = useForm<
+        Pick<AdditionalMetric, 'percentile'> & {
+            format: CustomFormat;
+            customMetricLabel: string;
+        }
+    >({
         validateInputOnChange: true,
         initialValues: {
             customMetricLabel: '',
@@ -132,7 +131,7 @@ export const CustomMetricModal = () => {
                     ? 'Metric with this label already exists'
                     : null;
             },
-            percentile: (percentile: number) => {
+            percentile: (percentile) => {
                 if (!percentile) return null;
                 if (percentile < 0 || percentile > 100) {
                     return 'Percentile must be a number between 0 and 100';
@@ -155,14 +154,6 @@ export const CustomMetricModal = () => {
                     : '',
             );
         }
-
-        if (
-            isEditing &&
-            isAdditionalMetric(item) &&
-            item.percentile !== undefined
-        ) {
-            setFieldValue('percentile', item.percentile);
-        }
     }, [setFieldValue, item, customMetricType, isEditing]);
 
     const initialCustomMetricFiltersWithIds = useMemo(() => {
@@ -184,6 +175,25 @@ export const CustomMetricModal = () => {
         setCustomMetricFiltersWithIds(initialCustomMetricFiltersWithIds);
     }, [initialCustomMetricFiltersWithIds]);
 
+    useEffect(
+        function populateForm() {
+            if (isEditing && isAdditionalMetric(item)) {
+                if (item.percentile)
+                    setFieldValue('percentile', item.percentile);
+
+                if (item.formatOptions) {
+                    setFieldValue('format', item.formatOptions);
+                }
+            }
+        },
+        [isEditing, item, setFieldValue],
+    );
+
+    const handleClose = useCallback(() => {
+        form.reset();
+        toggleModal();
+    }, [form, toggleModal]);
+
     const handleOnSubmit = form.onSubmit(
         ({ customMetricLabel, percentile, format }) => {
             if (!item || !customMetricType) return;
@@ -196,9 +206,7 @@ export const CustomMetricModal = () => {
                 isEditingCustomMetric: !!isEditing,
                 exploreData,
                 percentile,
-                ...(isCustomMetricFormattingEnabled && {
-                    customMetricFormat: format,
-                }),
+                formatOptions: format,
             });
 
             if (isEditing && isAdditionalMetric(item)) {
@@ -222,7 +230,7 @@ export const CustomMetricModal = () => {
                     title: 'Custom metric added successfully',
                 });
             }
-            toggleModal();
+            handleClose();
         },
     );
 
@@ -253,7 +261,7 @@ export const CustomMetricModal = () => {
             size="xl"
             onClick={(e) => e.stopPropagation()}
             opened={isOpen}
-            onClose={() => toggleModal(undefined)}
+            onClose={handleClose}
             title={
                 <Title order={4}>
                     {isEditing ? 'Edit' : 'Create'} Custom Metric